fix(db): widen interface counters to BIGINT; fix topology API 500
interface_stats and interface_ts declared in_octets/out_octets/prev_*
as INTEGER. Postgres treats that as int4 (signed 32-bit), which rejects
SNMP Counter64 values on real switches — busy 24h ports easily push past
2.1B bytes. Topology + monitoring writes were hitting "value out of int32
range" on every poll cycle.

Migration 0034 ALTERs the four interface_stats columns and two
interface_ts columns to BIGINT. SQLite uses variable-width INTEGER
affinity so it's a no-op there.

Also fix /api/topology returning 500: the route called
_db.execute_fetchall(), an aiosqlite shorthand that was never added to
the postgres adapter. Switched to the standard execute → fetchall
pattern (works on both engines) and added the missing _db.close() that
the original was leaking on every topology page load.

+++ routes/database.py
@@ -375,10 +375,13 @@ CREATE TABLE IF NOT EXISTS interface_stats (
     if_index            INTEGER NOT NULL,
     if_name             TEXT    NOT NULL DEFAULT '',
     if_speed_mbps       INTEGER DEFAULT 0,
-    in_octets           INTEGER DEFAULT 0,
-    out_octets          INTEGER DEFAULT 0,
-    prev_in_octets      INTEGER DEFAULT 0,
-    prev_out_octets     INTEGER DEFAULT 0,
+    -- Counter64 columns: real switches push tens of billions of bytes, which
+    -- overflow signed int32 (~2.1B). BIGINT in postgres / INTEGER affinity
+    -- in sqlite (which is variable-width) keeps both engines happy.
+    in_octets           BIGINT  DEFAULT 0,
+    out_octets          BIGINT  DEFAULT 0,
+    prev_in_octets      BIGINT  DEFAULT 0,
+    prev_out_octets     BIGINT  DEFAULT 0,
     polled_at           TEXT    NOT NULL DEFAULT (datetime('now')),
     prev_polled_at      TEXT    DEFAULT NULL,
     UNIQUE(host_id, if_index)
@@ -790,8 +793,8 @@ CREATE TABLE IF NOT EXISTS interface_ts (
     if_index        INTEGER NOT NULL,
     if_name         TEXT    NOT NULL DEFAULT '',
     if_speed_mbps   INTEGER DEFAULT 0,
-    in_octets       INTEGER DEFAULT 0,
-    out_octets      INTEGER DEFAULT 0,
+    in_octets       BIGINT  DEFAULT 0,
+    out_octets      BIGINT  DEFAULT 0,
     in_rate_bps     REAL    DEFAULT NULL,
     out_rate_bps    REAL    DEFAULT NULL,
     utilization_pct REAL    DEFAULT NULL,
